#10 Create project page and routes
