+++ src/components/project-page/index.html
@@ -1,0 +1,134 @@
+<div id="readme" class="md" data-path="README.md">
+	<article class="markdown-body entry-content container-lg" itemprop="text">
+		<p dir="auto">
+			<a href="https://github.com/wwcodecolorado/scrollodex/blob/master/LICENSE"
+				><img
+					src="https://camo.githubusercontent.com/34793017700039d2a040b5bfbb45438e66e88e33e8049cc8ac7c8939e68f86e0/68747470733a2f2f696d672e736869656c64732e696f2f6769746875622f6c6963656e73652f7777636f6465636f6c6f7261646f2f7363726f6c6c6f646578"
+					alt="GitHub License"
+					data-canonical-src="https://img.shields.io/github/license/wwcodecolorado/scrollodex"
+					style="max-width: 100%" /></a
+			><a href="https://app.netlify.com/sites/scrollodex/deploys" rel="nofollow"
+				><img
+					src="https://camo.githubusercontent.com/99ed14391c42025adec3b5b5ee9444b790ade721d33d440da5534f14f6dfe652/68747470733a2f2f6170692e6e65746c6966792e636f6d2f6170692f76312f6261646765732f39623732613134332d393365362d346165322d613036352d3562373263613537623637362f6465706c6f792d737461747573"
+					alt="Netlify Status"
+					data-canonical-src="https://api.netlify.com/api/v1/badges/9b72a143-93e6-4ae2-a065-5b72ca57b676/deploy-status"
+					style="max-width: 100%"
+			/></a>
+		</p>
+		<h1 dir="auto">
+			<a
+				id="user-content-scrollodex"
+				class="anchor"
+				aria-hidden="true"
+				href="#scrollodex"
+				><svg
+					class="octicon octicon-link"
+					viewBox="0 0 16 16"
+					version="1.1"
+					width="16"
+					height="16"
+					aria-hidden="true"
+				>
+					<path
+						fill-rule="evenodd"
+						d="M7.775 3.275a.75.75 0 001.06 1.06l1.25-1.25a2 2 0 112.83 2.83l-2.5 2.5a2 2 0 01-2.83 0 .75.75 0 00-1.06 1.06 3.5 3.5 0 004.95 0l2.5-2.5a3.5 3.5 0 00-4.95-4.95l-1.25 1.25zm-4.69 9.64a2 2 0 010-2.83l2.5-2.5a2 2 0 012.83 0 .75.75 0 001.06-1.06 3.5 3.5 0 00-4.95 0l-2.5 2.5a3.5 3.5 0 004.95 4.95l1.25-1.25a.75.75 0 00-1.06-1.06l-1.25 1.25a2 2 0 01-2.83 0z"
+					></path></svg></a
+			>Scrollodex
+		</h1>
+		<p dir="auto">
+			Welcome! Scrollodex is an open-source project created and maintained by
+			members of the Women Who Code Boulder/Denver in order to bring the
+			community together, build something awesome, and spend some time hacking!
+			Based on the classic rolodex or a business card pin wall, Scrollodex aims
+			to be a fun place for Women Who Code members to add and display a business
+			card.
+		</p>
+		<h2 dir="auto">
+			<a
+				id="user-content-contributing"
+				class="anchor"
+				aria-hidden="true"
+				href="#contributing"
+				><svg
+					class="octicon octicon-link"
+					viewBox="0 0 16 16"
+					version="1.1"
+					width="16"
+					height="16"
+					aria-hidden="true"
+				>
+					<path
+						fill-rule="evenodd"
+						d="M7.775 3.275a.75.75 0 001.06 1.06l1.25-1.25a2 2 0 112.83 2.83l-2.5 2.5a2 2 0 01-2.83 0 .75.75 0 00-1.06 1.06 3.5 3.5 0 004.95 0l2.5-2.5a3.5 3.5 0 00-4.95-4.95l-1.25 1.25zm-4.69 9.64a2 2 0 010-2.83l2.5-2.5a2 2 0 012.83 0 .75.75 0 001.06-1.06 3.5 3.5 0 00-4.95 0l-2.5 2.5a3.5 3.5 0 004.95 4.95l1.25-1.25a.75.75 0 00-1.06-1.06l-1.25 1.25a2 2 0 01-2.83 0z"
+					></path></svg></a
+			>Contributing
+		</h2>
+		<p dir="auto">
+			Our project is ready to accept PRs! We launched it on Oct 13, 2021 in our
+			<a
+				href="https://www.meetup.com/Women-Who-Code-Boulder-Denver/events/281109143/"
+				rel="nofollow"
+				>Meetup</a
+			>.
+			<a href="CONTRIBUTING.md">See our full contribution guidelines.</a> Check
+			out our
+			<a href="https://github.com/wwcodecolorado/scrollodex/issues">issues</a>
+			page to see what you can work on , we recommend issue #4!
+		</p>
+		<p dir="auto">
+			To get started, read and follow along with our
+			<a href="HOWTO101.md">How To 101: The Non-Scary Guide</a>!
+		</p>
+		<h2 dir="auto">
+			<a
+				id="user-content-requirements"
+				class="anchor"
+				aria-hidden="true"
+				href="#requirements"
+				><svg
+					class="octicon octicon-link"
+					viewBox="0 0 16 16"
+					version="1.1"
+					width="16"
+					height="16"
+					aria-hidden="true"
+				>
+					<path
+						fill-rule="evenodd"
+						d="M7.775 3.275a.75.75 0 001.06 1.06l1.25-1.25a2 2 0 112.83 2.83l-2.5 2.5a2 2 0 01-2.83 0 .75.75 0 00-1.06 1.06 3.5 3.5 0 004.95 0l2.5-2.5a3.5 3.5 0 00-4.95-4.95l-1.25 1.25zm-4.69 9.64a2 2 0 010-2.83l2.5-2.5a2 2 0 012.83 0 .75.75 0 001.06-1.06 3.5 3.5 0 00-4.95 0l-2.5 2.5a3.5 3.5 0 004.95 4.95l1.25-1.25a.75.75 0 00-1.06-1.06l-1.25 1.25a2 2 0 01-2.83 0z"
+					></path></svg></a
+			>Requirements
+		</h2>
+		<ul dir="auto">
+			<li>Node 16+</li>
+			<li>Yarn</li>
+		</ul>
+		<h2 dir="auto">
+			<a
+				id="user-content-installation"
+				class="anchor"
+				aria-hidden="true"
+				href="#installation"
+				><svg
+					class="octicon octicon-link"
+					viewBox="0 0 16 16"
+					version="1.1"
+					width="16"
+					height="16"
+					aria-hidden="true"
+				>
+					<path
+						fill-rule="evenodd"
+						d="M7.775 3.275a.75.75 0 001.06 1.06l1.25-1.25a2 2 0 112.83 2.83l-2.5 2.5a2 2 0 01-2.83 0 .75.75 0 00-1.06 1.06 3.5 3.5 0 004.95 0l2.5-2.5a3.5 3.5 0 00-4.95-4.95l-1.25 1.25zm-4.69 9.64a2 2 0 010-2.83l2.5-2.5a2 2 0 012.83 0 .75.75 0 001.06-1.06 3.5 3.5 0 00-4.95 0l-2.5 2.5a3.5 3.5 0 004.95 4.95l1.25-1.25a.75.75 0 00-1.06-1.06l-1.25 1.25a2 2 0 01-2.83 0z"
+					></path></svg></a
+			>Installation
+		</h2>
+		<p dir="auto">In the project directory, you can run:</p>
+		<p dir="auto"><code>yarn install</code><br /><code>yarn start</code></p>
+		<p dir="auto">
+			This runs the app in the development mode. Open
+			<a href="http://localhost:3000" rel="nofollow">http://localhost:3000</a>
+			to view it in the browser. The page will reload if you make edits.
+		</p>
+	</article>
+</div>
