+++ src/routes/home/home.component.jsx
@@ -4,7 +4,7 @@ import 'simplebar/dist/simplebar.min.css';
 import './home.styles.scss';
 import KatyaHeadshot from '../../assets/KatyaHeadshot.jpg';
 
-const Home = (test) => {
+const Home = () => {
     const codes = {
         leftBrace: 123,
         rightBrace: 125,
@@ -18,6 +18,7 @@ const Home = (test) => {
 
     return (
         <SimpleBar className="about-container">
+            {/* eslint-disable-next-line */}
             <p className="comment intro">// HI, I'M KATYA SARMIENTO</p><br />
             <p>
                 <span className="code">import</span> <span className="style">'./professional.styles.scss'</span>;
